libckteec: serialize: allow to use 4 bytes buffer for CK_ULONG attributes
PKCS11 TA uses uint32_t for CK_ULONG attributes and returns 4 bytes size
if length of attributes is requested. But libckteec returns error if
client get/set the attribute with 4 bytes buffer.

In current implementation, the length of client buffer is compared with
CK_ULONG size (which is defined as 8 bytes length in libcktee). Then the
buffer copied to CK_ULONG local variable and casts to uint32_t.

This commit checks if client buffer is at least 4 bytes long and copies
only lower 4 bytes of the client buffer. It is the same as current
implementation except the client buffer length is compared to size of
uint32_t.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>